refactor: moved auth routes to handler
